basic tests as well as changed frame allocater to 64 bit return and finished up mapping vaddr to paddr

+++ src/frame.cpp
@@ -25,12 +25,12 @@ void create_frame_table(uintptr_t start, int size) {
     }
 }
 
-void alloc_frame(int flags, Function<void(int)> w) {
+void alloc_frame(int flags, Function<void(uint64_t)> w) {
     for (int i = 0; i < num_frames; i++) {
         if (!(frame_table[index].flags & USED_PAGE_FLAG)){
             frame_table[index].flags = flags;
             frame_table[index].flags |= USED_PAGE_FLAG;
-            create_event_value<int>(w, index * PAGE_SIZE, 1);
+            create_event_value<uint64_t>(w, index * PAGE_SIZE, 1);
             index += 1;
             return;
         }
